Готова интерполяция

+++ src/main/java/model/Function2Args.java
@@ -5,9 +5,9 @@ package model;
  */
 public class Function2Args {
     public final Coordinate coordinate;
-    public final int value;
+    public final double value;
 
-    public Function2Args(int x, int y, int value) {
+    public Function2Args(int x, int y, double value) {
         coordinate = new Coordinate(x, y);
         this.value = value;
     }
